# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hush Service API
  version: 1.0.0
  description: Single-owner encrypted vault. CLI devices are read-only. All binary
    fields use canonical unpadded base64url. See crypto.md for mandatory AAD,
    algorithm, size, and nonce rules. The Worker validates structure, not
    cryptographic correctness.
servers:
  - url: /api/v1
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      summary: List profiles and environment mappings
      security: *a4
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Profile"
                required:
                  - profiles
                additionalProperties: false
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /profiles/{id}:
    get:
      operationId: getProfile
      summary: Read one profile
      security: *a4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: putProfile
      summary: Create or replace profile mappings
      security: *a3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
    delete:
      operationId: deleteProfile
      summary: Delete a profile
      security: *a3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /profiles/{id}/secrets:
    get:
      operationId: getProfileSecrets
      summary: Fetch only ciphertext selected by a profile
      security: *a4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: "#/components/schemas/Profile"
                  secrets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Secret"
                required:
                  - profile
                  - secrets
                additionalProperties: false
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    ProfileInput:
      type: object
      properties:
        name: &a11
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[^\x00-\x1f\x7f]+$
        mappings: &a12
          maxItems: 500
          type: array
          items:
            type: object
            properties:
              secret_id:
                type: string
                format: uuid
                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
              env_name:
                type: string
                minLength: 1
                maxLength: 128
                pattern: ^[A-Za-z_][A-Za-z0-9_]*$
            required:
              - secret_id
              - env_name
            additionalProperties: false
      required:
        - name
        - mappings
      additionalProperties: false
...
    Secret:
      type: object
      properties:
        name: *a6
        version: *a7
        vek_version: *a8
        envelope: *a9
        id: &a24
          type: string
          format: uuid
        created_at: &a10
          type: string
          format: date-time
        updated_at: *a10
      required:
        - name
        - version
        - vek_version
        - envelope
        - id
        - created_at
        - updated_at
      additionalProperties: false
    Profile:
      type: object
      properties:
        name: *a11
        mappings: *a12
        id:
          type: string
          format: uuid
        created_at: *a10
        updated_at: *a10
      required:
        - name
        - mappings
        - id
        - created_at
        - updated_at
      additionalProperties: false
...
    Revision:
      type: object
      properties:
        revision: *a23
      required:
        - revision
      additionalProperties: false
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *a2
            request_id: *a2
          required:
            - code
            - request_id
          additionalProperties: false
      required:
        - error
      additionalProperties: false
  securitySchemes:
    AccessSession:
      type: apiKey
      in: cookie
      name: CF_Authorization
      description: Cloudflare Access human session. Edge forwards the signed
        assertion; Worker verifies it and matches OWNER_EMAIL.
    AccessClientId:
      type: apiKey
      in: header
      name: CF-Access-Client-Id
    AccessClientSecret:
      type: apiKey
      in: header
      name: CF-Access-Client-Secret